Alle Rechte drinne, bis auf getArbeitsgruppeDromMitarbeiter

+++ Backend/src/main/java/com/nsj/gfos/award/resources/ArbeitsgruppenResource.java
@@ -1,6 +1,5 @@
 package com.nsj.gfos.award.resources;
 
-import javax.swing.text.html.HTMLDocument.HTMLReader.IsindexAction;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
@@ -9,6 +8,7 @@ import javax.ws.rs.core.MediaType;
 
 import com.nsj.gfos.award.handlers.JsonHandler;
 import com.nsj.gfos.award.handlers.QueryHandler;
+import com.nsj.gfos.award.handlers.RightHandler;
 import com.nsj.gfos.award.handlers.SessionHandler;
 import com.nsj.gfos.award.gUtils.Utils;
 
@@ -44,6 +44,7 @@ public class ArbeitsgruppenResource {
 		String pn = attributes[1].split("=")[1];
 		if (pn.length() != 12)
 			return JsonHandler.fehler("Ungültige Personalnummer.");
+		
 		// TODO checkRights
 		if (!SessionHandler.checkSessionID(auth))
 			return JsonHandler.fehler("SessionID ist ungültig.");
@@ -57,7 +58,11 @@ public class ArbeitsgruppenResource {
 			if (!rs.next())
 				return JsonHandler.fehler("Leere Rückgabe der Datenbank.");
 			Arbeitsgruppe a = Utils.createArbeitsgruppeFromQuery(rs);
-			return JsonHandler.createJsonFromArbeitsgruppe(a);
+			if (Utils.isInArbeitsgruppe(a.getArbeitsgruppenID(), Utils.getPnFromSessionID(auth))
+					|| RightHandler.checkPermission(auth, "getArbeitsgruppeFromMitarbeiter")) {
+				return JsonHandler.createJsonFromArbeitsgruppe(a);
+			}
+			return JsonHandler.fehler("Der Mitarbeiter hat keine Berechtigung.");
 		} catch (SQLException e) {
 			return JsonHandler.fehler(e.toString());
 		}
@@ -92,24 +97,17 @@ public class ArbeitsgruppenResource {
 			return JsonHandler.fehler("Ungültige Personalnummer.");
 		if (id.length() != 12)
 			return JsonHandler.fehler("Ungültige ArbeitsgruppenID.");
-		// TODO checkRights + die des Leiters
 		if (!SessionHandler.checkSessionID(auth))
 			return JsonHandler.fehler("SessionID ist ungültig.");
 		if (!Utils.isInArbeitsgruppe(id, pn))
 			return JsonHandler.fehler("Der Leiter muss bereits in der Arbeitsgruppe sein.");
-		String sqlStmt = "SELECT gfos.arbeitsgruppe.Leiter FROM gfos.Arbeitsgruppe WHERE gfos.arbeitsgruppe.ArbeitsgruppenID = \""
-				+ id + "\";";
-		try {
-			ResultSet rs = QueryHandler.query(sqlStmt);
-			if (rs == null)
-				return JsonHandler.fehler("Der Leiter der Arbeitsgruppe konnte nicht gefunden werden.");
-			rs.next();
-			if (id.equals(rs.getString("Leiter")))
-				return JsonHandler.fehler("Der Mitarbeiter ist bereits der Leiter der Arbeitsgruppe.");
-		} catch (SQLException e) {
-			return JsonHandler.fehler(e.toString());
-		}
-		sqlStmt = "UPDATE gfos.arbeitsgruppe SET gfos.arbeitsgruppe.Leiter = \"" + pn
+		if(Utils.getLeiter(id).equals(pn))
+			return JsonHandler.fehler("Der Mitarbeiter ist bereits der Leiter der Arbeitsgruppe.");
+		if(!RightHandler.checkPermissionFromPn(pn, "becomeLeiter"))
+			return JsonHandler.fehler("Der angegebene Leiter hat keine Berechtigung Leiter zu werden.");
+		if(!RightHandler.checkPermission(auth, "alterLeiter"))
+			return JsonHandler.fehler("Der Mitarbeiter hat keine Berechtigung.");
+		String sqlStmt = "UPDATE gfos.arbeitsgruppe SET gfos.arbeitsgruppe.Leiter = \"" + pn
 				+ "\" WHERE gfos.arbeitsgruppe.ArbeitsgruppenID = \"" + id + "\";";
 		try {
 			int rs = QueryHandler.update(sqlStmt);
@@ -153,7 +151,10 @@ public class ArbeitsgruppenResource {
 			return JsonHandler.fehler("Der Mitarbeiter, der Leiter werden soll, existiert nicht.");
 		if (!SessionHandler.checkSessionID(auth))
 			return JsonHandler.fehler("SessionID ist ungültig.");
-		// TODO checkRights + die des Leiters
+		if(!RightHandler.checkPermission(auth, "addArbeitsgruppe"))
+			return JsonHandler.fehler("Der Mitarbeiter hat keine Berechtigung.");
+		if(!RightHandler.checkPermissionFromPn(leiter, "becomeLeiter"))
+			return JsonHandler.fehler("Der angegebene Leiter hat keine Berechtigung Leiter zu werden.");
 		String arbeitsgruppenID = Utils.createArbeitsgruppenID();
 		String sqlStmt = "INSERT INTO gfos.arbeitsgruppe (ArbeitsgruppenID, Bezeichnung, Leiter) Values ('"
 				+ arbeitsgruppenID + "', '" + bezeichnung + "', '" + leiter + "');";
@@ -205,7 +206,9 @@ public class ArbeitsgruppenResource {
 			return JsonHandler.fehler("SessionID ist ungültig.");
 		if (!Utils.checkIfArbeitsgruppeExistsFromID(id))
 			return JsonHandler.fehler("Arbeitsgruppe existiert nicht.");
-		// TODO checkRights
+		if (!((Utils.isInArbeitsgruppe(id, Utils.getPnFromSessionID(auth)) && Utils.getLeiter(id).equals(Utils.getPnFromSessionID(auth)))
+				|| RightHandler.checkPermission(auth, "removeArbeitsgruppe")))
+			return JsonHandler.fehler("Der Mitarbeiter hat keine Berechtigung.");
 		String sqlStmt = "DELETE FROM gfos.arbeitsgruppenteilnahme WHERE gfos.arbeitsgruppenteilnahme.ArbeitsgruppenID = \""
 				+ id + "\";";
 		try {
@@ -256,24 +259,16 @@ public class ArbeitsgruppenResource {
 			return JsonHandler.fehler("Ungültige Personalnummer.");
 		if (id.length() != 12)
 			return JsonHandler.fehler("Ungültige ArbeitsgruppenID.");
-		if (!Utils.isInArbeitsgruppe(id, pn))
-			return JsonHandler.fehler("Der Mitarbeiter befindet sich nicht in der Arbeitsgruppe.");
-		// TODO checkRights
 		if (!SessionHandler.checkSessionID(auth))
 			return JsonHandler.fehler("SessionID ist ungültig.");
-		String sqlStmt = "SELECT gfos.arbeitsgruppe.Leiter FROM gfos.Arbeitsgruppe WHERE gfos.arbeitsgruppe.ArbeitsgruppenID = \""
-				+ id + "\";";
-		try {
-			ResultSet rs = QueryHandler.query(sqlStmt);
-			if (rs == null)
-				return JsonHandler.fehler("Der Leiter der Arbeitsgruppe konnte nicht gefunden werden.");
-			rs.next();
-			if (id.equals(rs.getString("Leiter")))
-				return JsonHandler.fehler("Der Leiter einer Arbeitsgruppe kann nicht gelöscht werden.");
-		} catch (SQLException e) {
-			return JsonHandler.fehler(e.toString());
-		}
-		sqlStmt = "DELETE FROM gfos.arbeitsgruppenteilnahme WHERE gfos.arbeitsgruppenteilnahme.ArbeitsgruppenID = \""
+		if (!Utils.isInArbeitsgruppe(id, pn))
+			return JsonHandler.fehler("Der Mitarbeiter befindet sich nicht in der Arbeitsgruppe.");
+		if (Utils.getLeiter(id).equals(pn))
+			return JsonHandler.fehler("Der Leiter der Arbeitsgruppe kann nicht gelöscht werden.");
+		if (!((Utils.isInArbeitsgruppe(id, Utils.getPnFromSessionID(auth)) && Utils.getLeiter(id).equals(Utils.getPnFromSessionID(auth)))
+				|| RightHandler.checkPermission(auth, "removeMitarbeiterFromArbeitsgruppe")))
+			return JsonHandler.fehler("Der Mitarbeiter hat keine Berechtigung.");
+		String sqlStmt = "DELETE FROM gfos.arbeitsgruppenteilnahme WHERE gfos.arbeitsgruppenteilnahme.ArbeitsgruppenID = \""
 				+ id + "\" AND gfos.arbeitsgruppenteilnahme.Mitarbeiter = \"" + pn + "\";";
 		try {
 			int rs = QueryHandler.update(sqlStmt);
@@ -310,20 +305,23 @@ public class ArbeitsgruppenResource {
 		}
 		String auth = attributes[0].split("=")[1];
 		String pn = attributes[1].split("=")[1];
-		String arbeitsgruppenID = attributes[2].split("=")[1];
+		String id = attributes[2].split("=")[1];
 		if (pn.length() != 12)
 			return JsonHandler.fehler("Ungültige Personalnummer.");
-		if (!Utils.checkIfArbeitsgruppeExistsFromID(arbeitsgruppenID))
+		if (!SessionHandler.checkSessionID(auth))
+			return JsonHandler.fehler("SessionID ist ungültig.");
+		if (!Utils.checkIfArbeitsgruppeExistsFromID(id))
 			return JsonHandler.fehler("Arbeitsgruppe existiert nicht.");
 		if (!Utils.checkIfMitarbeiterExists(pn))
 			return JsonHandler.fehler("Der Mitarbeiter existiert nicht.");
-		if (Utils.isInArbeitsgruppe(arbeitsgruppenID, pn))
+		if (Utils.isInArbeitsgruppe(id, pn))
 			return JsonHandler.fehler("Der Mitarbeiter ist bereits in der Arbeitsgruppe vorhanden.");
-		if (!SessionHandler.checkSessionID(auth))
-			return JsonHandler.fehler("SessionID ist ungültig.");
-		// TODO checkRights
+		if (!((Utils.isInArbeitsgruppe(id, Utils.getPnFromSessionID(auth))
+				&& Utils.getLeiter(id).equals(Utils.getPnFromSessionID(auth)))
+				|| RightHandler.checkPermission(auth, "addMitarbeiterToArbeitsgruppe")))
+			return JsonHandler.fehler("Der Mitarbeiter hat keine Berechtigung.");
 		String sqlStmt = "INSERT INTO gfos.arbeitsgruppenteilnahme (ArbeitsgruppenID, Mitarbeiter) Values ('"
-				+ arbeitsgruppenID + "', '" + pn + "');";
+				+ id + "', '" + pn + "');";
 		try {
 			int rs = QueryHandler.update(sqlStmt);
 			if (rs == 0)
